// frontend/src/app/[locale]/(auth)/register/page.tsx
'use client';

import { zodResolver } from '@hookform/resolvers/zod';
import {
  ArrowRight,
  CalendarIcon,
  ChevronRight,
  Eye,
  EyeOff,
  Sparkles,
} from 'lucide-react';
import Image from 'next/image';
import Link from 'next/link';
import { useRouter } from 'next/navigation';
import { useState } from 'react';
import { Controller, useForm } from 'react-hook-form';
import * as z from 'zod';

import {
  Button,
  Calendar,
  Card,
  CardContent,
  CardDescription,
  CardFooter,
  CardHeader,
  CardTitle,
  Checkbox,
  Field,
  FieldError,
  FieldGroup,
  FieldLabel,
  Input,
  InputGroup,
  InputGroupButton,
  InputGroupInput,
  Popover,
  PopoverContent,
  PopoverTrigger,
  Spinner,
} from '@/components/ui';
import { RegisterAuthDto } from '@/generated/model';
import { useRegister } from '@/hooks/api/useAuth';
import { toApiDateString } from '@/utils';

const registerFormSchema = z
  .object({
    email: z.email(),
    password: z
      .string()
      .min(6, 'Mật khẩu phải có ít nhất 6 ký tự.')
      .max(50, 'Mật khẩu không được vượt quá 50 ký tự.')
      .regex(
        /^(?=.*[a-z])(?=.*[A-Z])(?=.*\d)/,
        'Mật khẩu phải chứa ít nhất một chữ cái viết hoa, một chữ cái viết thường và một chữ số.',
      ),
    confirmPassword: z.string().trim().min(1, 'Xác nhận mật khẩu là bắt buộc.'),
    firstName: z.string().trim().min(1, 'Tên là bắt buộc.'),
    lastName: z.string().trim().min(1, 'Họ là bắt buộc.'),
    phone: z.string().trim().optional(),
    dateOfBirth: z.date().optional(),
  })
  .refine((data) => data.password === data.confirmPassword, {
    message: 'Mật khẩu xác nhận không khớp',
    path: ['confirmPassword'],
  });

// FIXME: Logged in user is not redirected to dashboard page, but stays on register page
export default function RegisterPage() {
  const router = useRouter();
  const registerMutation = useRegister();
  const form = useForm<z.infer<typeof registerFormSchema>>({
    resolver: zodResolver(registerFormSchema),
    defaultValues: {
      email: '',
      password: '',
      confirmPassword: '',
      firstName: '',
      lastName: '',
      phone: '',
    },
  });

  const [agreeTerms, setAgreeTerms] = useState(false);
  const [showPassword, setShowPassword] = useState(false);
  const [popoverIsOpen, setPopoverIsOpen] = useState(false);

  const onSubmit = (data: z.infer<typeof registerFormSchema>) => {
    // eslint-disable-next-line @typescript-eslint/no-unused-vars
    const { confirmPassword, dateOfBirth, ...rest } = data;
    const registerAuthDto = {
      ...rest,
      ...(dateOfBirth ? { dateOfBirth: toApiDateString(dateOfBirth) } : {}),
    } as RegisterAuthDto;

    registerMutation.mutate(registerAuthDto, {
      onSuccess: () => {
        router.push('/login?registered=true');
      },
    });
  };

  return (
    <div className="flex h-screen w-full overflow-hidden">
      {/* Left: Hero image */}
      <div className="relative hidden flex-1 overflow-hidden lg:block">
        <Image
          src="/images/register-hero-1.png"
          alt="register hero image"
          fill
          sizes="50vw"
          priority
          className="object-cover"
        />
        <div className="from-primary/70 to-primary/0 absolute inset-0 bg-linear-to-r" />
        <div className="absolute bottom-12 left-12 flex max-w-xl flex-col gap-4">
          <h1 className="text-4xl leading-tight font-bold tracking-tight text-white">
            Bắt đầu với TacoLearn.
          </h1>
          <p className="text-base text-white/90">
            Tạo tài khoản để bắt đầu sử dụng nền tảng.
          </p>
        </div>
      </div>

      {/* Right: Register form */}
      <div className="flex h-full flex-1 flex-col overflow-y-auto bg-white">
        <div className="sticky top-0 z-10 flex items-center justify-between border-b border-gray-100 bg-white/90 px-6 py-4 backdrop-blur-sm">
          {/* TODO: create logo component for web */}
          <Link href="/" className="flex items-center gap-2">
            <div className="flex size-8 items-center justify-center rounded-lg bg-indigo-600">
              <Sparkles className="size-4 text-white" />
            </div>
            <span className="text-lg font-semibold text-gray-900">
              TacoLearn
            </span>
          </Link>
          <Link
            href="/login"
            className="flex items-center gap-1 text-sm font-medium text-gray-600 hover:text-gray-900"
          >
            Đăng nhập
            <ChevronRight className="size-3" />
          </Link>
        </div>

        <div className="flex flex-1 items-center justify-center-safe px-6 py-10">
          <Card className="w-full max-w-lg ring-0">
            <CardHeader>
              <CardTitle>
                <h2 className="text-4xl font-bold tracking-tight text-gray-900">
                  Tạo tài khoản
                </h2>
              </CardTitle>
              <CardDescription>
                <p className="text-sm font-medium text-gray-600">
                  Điền thông tin để đăng ký tài khoản của bạn.
                </p>
              </CardDescription>
            </CardHeader>
            <CardContent>
              <form id="register-form" onSubmit={form.handleSubmit(onSubmit)}>
                <FieldGroup>
                  <div className="grid grid-cols-2 gap-3">
                    <Controller
                      name="lastName"
                      control={form.control}
                      render={({ field, fieldState }) => (
                        <Field data-invalid={fieldState.invalid}>
                          <FieldLabel htmlFor="lastName">Họ *</FieldLabel>
                          <Input
                            {...field}
                            id="lastName"
                            aria-invalid={fieldState.invalid}
                            placeholder="Nguyễn"
                          />
                          {fieldState.invalid && (
                            <FieldError errors={[fieldState.error]} />
                          )}
                        </Field>
                      )}
                    />
                    <Controller
                      name="firstName"
                      control={form.control}
                      render={({ field, fieldState }) => (
                        <Field data-invalid={fieldState.invalid}>
                          <FieldLabel htmlFor="firstName">Tên *</FieldLabel>
                          <Input
                            {...field}
                            id="firstName"
                            aria-invalid={fieldState.invalid}
                            placeholder="Văn A"
                          />
                          {fieldState.invalid && (
                            <FieldError errors={[fieldState.error]} />
                          )}
                        </Field>
                      )}
                    />
                  </div>

                  <Controller
                    name="email"
                    control={form.control}
                    render={({ field, fieldState }) => (
                      <Field data-invalid={fieldState.invalid}>
                        <FieldLabel htmlFor="email">Email *</FieldLabel>
                        <Input
                          {...field}
                          id="email"
                          aria-invalid={fieldState.invalid}
                          placeholder="email@example.com"
                        />
                        {fieldState.invalid && (
                          <FieldError errors={[fieldState.error]} />
                        )}
                      </Field>
                    )}
                  />

                  <Controller
                    name="phone"
                    control={form.control}
                    render={({ field, fieldState }) => (
                      <Field data-invalid={fieldState.invalid}>
                        <FieldLabel htmlFor="phone">Số điện thoại</FieldLabel>
                        <Input
                          {...field}
                          id="phone"
                          aria-invalid={fieldState.invalid}
                          placeholder="0901234567"
                        />
                        {fieldState.invalid && (
                          <FieldError errors={[fieldState.error]} />
                        )}
                      </Field>
                    )}
                  />

                  <Controller
                    name="dateOfBirth"
                    control={form.control}
                    render={({ field, fieldState }) => (
                      <Field data-invalid={fieldState.invalid}>
                        <FieldLabel htmlFor="dateOfBirth">Ngày sinh</FieldLabel>
                        <Popover
                          open={popoverIsOpen}
                          onOpenChange={setPopoverIsOpen}
                        >
                          <PopoverTrigger asChild>
                            <Button
                              id="dateOfBirth"
                              type="button"
                              variant="outline"
                              className="w-full justify-start font-normal"
                            >
                              <CalendarIcon />
                              {field.value ? (
                                field.value.toLocaleDateString()
                              ) : (
                                <span className="text-gray-500">
                                  {'dd/mm/yyyy'}
                                </span>
                              )}
                            </Button>
                          </PopoverTrigger>
                          <PopoverContent>
                            <Calendar
                              mode="single"
                              selected={field.value}
                              defaultMonth={field.value}
                              captionLayout="dropdown"
                              onSelect={(date) => {
                                field.onChange(date);
                                setPopoverIsOpen(false);
                              }}
                            />
                          </PopoverContent>
                        </Popover>
                        {fieldState.invalid && (
                          <FieldError errors={[fieldState.error]} />
                        )}
                      </Field>
                    )}
                  />

                  <div className="grid grid-cols-2 gap-3">
                    <Controller
                      name="password"
                      control={form.control}
                      render={({ field, fieldState }) => (
                        <Field data-invalid={fieldState.invalid}>
                          <FieldLabel htmlFor="password">Mật khẩu *</FieldLabel>
                          <InputGroup>
                            <InputGroupInput
                              {...field}
                              id="password"
                              type={showPassword ? 'text' : 'password'}
                              aria-invalid={fieldState.invalid}
                              placeholder="••••••••"
                            />
                            <InputGroupButton
                              variant="ghost"
                              onClick={() => setShowPassword((prev) => !prev)}
                              className="h-full text-gray-400 hover:text-gray-600"
                            >
                              {showPassword ? (
                                <EyeOff className="size-4" />
                              ) : (
                                <Eye className="size-4" />
                              )}
                            </InputGroupButton>
                          </InputGroup>
                          {fieldState.invalid && (
                            <FieldError errors={[fieldState.error]} />
                          )}
                        </Field>
                      )}
                    />
                    <Controller
                      name="confirmPassword"
                      control={form.control}
                      render={({ field, fieldState }) => (
                        <Field data-invalid={fieldState.invalid}>
                          <FieldLabel htmlFor="confirmPassword">
                            Xác nhận mật khẩu *
                          </FieldLabel>
                          <InputGroup>
                            <InputGroupInput
                              {...field}
                              id="confirmPassword"
                              type={showPassword ? 'text' : 'password'}
                              aria-invalid={fieldState.invalid}
                              placeholder="••••••••"
                            />
                            <InputGroupButton
                              variant="ghost"
                              onClick={() => setShowPassword((prev) => !prev)}
                              className="h-full text-gray-400 hover:text-gray-600"
                            >
                              {showPassword ? (
                                <EyeOff className="size-4" />
                              ) : (
                                <Eye className="size-4" />
                              )}
                            </InputGroupButton>
                          </InputGroup>
                          {fieldState.invalid && (
                            <FieldError errors={[fieldState.error]} />
                          )}
                        </Field>
                      )}
                    />
                  </div>

                  <Field orientation="horizontal" className="relative">
                    <Checkbox
                      id="agreeTerms"
                      checked={agreeTerms}
                      onCheckedChange={() => setAgreeTerms((prev) => !prev)}
                    />
                    <FieldLabel htmlFor="agreeTerms" className="text-gray-700">
                      <span>
                        Tôi đồng ý với{' '}
                        <span className="text-indigo-600">
                          Điều khoản dịch vụ
                        </span>{' '}
                        và{' '}
                        <span className="text-indigo-600">
                          Chính sách bảo mật
                        </span>
                        .
                      </span>
                    </FieldLabel>
                  </Field>
                </FieldGroup>
              </form>
            </CardContent>
            <CardFooter>
              <Button
                type="submit"
                form="register-form"
                className="h-12 w-full bg-indigo-600 text-sm font-medium shadow-sm hover:bg-indigo-700"
                disabled={registerMutation.isPending || !agreeTerms}
              >
                {registerMutation.isPending ? (
                  <Spinner />
                ) : (
                  <>
                    Tạo tài khoản
                    <ArrowRight className="size-3" />
                  </>
                )}
              </Button>
            </CardFooter>
          </Card>
        </div>
      </div>
    </div>
  );
}
